rag_008_graphrag_neo4j_ontology/requirements.txt
python-dotenv>=1.0.1
pyyaml>=6.0.2
langchain>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-neo4j>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0
neo4j>=5.23.0
pyvis>=0.3.2
pytest>=8.3.0
